Adding Google Test support.
This is a slightly older version of the gtest library.  Will be updating to the latest gtest-1.5 release next.

A developer needs only add a test to the source code, and re-run CMake to have ctest pick up the new test(s).  Plese reference http://code.google.com/p/googletest/ for examples and documentation.